Cleaned up code and fixing minor bugs.
